+++ src/utils.js
@@ -1,5 +1,6 @@
 import contentTypeParser from 'content-type';
 import os from 'os';
+import _ from 'underscore';
 import fs from 'fs';
 import util from 'util';
 import fsExtra from 'fs-extra';
@@ -8,13 +9,15 @@ import psTree from '@apify/ps-tree';
 import requestPromise from 'request-promise-native';
 import XRegExp from 'xregexp';
 import cheerio from 'cheerio';
+import log from 'apify-shared/log';
 import { delayPromise, getRandomInt } from 'apify-shared/utilities';
 import { ENV_VARS, LOCAL_ENV_VARS } from 'apify-shared/consts';
 import { checkParamOrThrow } from 'apify-client/build/utils';
+import { version as apifyClientVersion } from 'apify-client/package.json';
+import { version as apifyVersion } from '../package.json';
 import { USER_AGENT_LIST } from './constants';
 
 /* globals process */
-
 
 /**
  * Default regular expression to match URLs in a string that may be plain text, JSON, CSV or other. It supports common URL characters
@@ -51,6 +54,18 @@ export const newClient = () => {
     if (apiBaseUrl) opts.baseUrl = apiBaseUrl;
 
     return new ApifyClient(opts);
+};
+
+/**
+ * Logs info about system, node version and apify package version.
+ */
+export const logSystemInfo = () => {
+    log.info('System info', {
+        apifyVersion,
+        apifyClientVersion,
+        osType: os.type(),
+        nodeVersion: process.version,
+    });
 };
 
 /**
@@ -560,6 +575,35 @@ const htmlToText = (html) => {
     return text.trim();
 };
 
+/**
+ * Creates a standardized debug info from request and response. This info is usually added to dataset under the hidden `#debug` field.
+ *
+ * @param {Object} request [Apify.Request](https://sdk.apify.com/docs/api/request) object.
+ * @param {Object} [response] Puppeteer [Response](https://pptr.dev/#?product=Puppeteer&version=v1.11.0&show=api-class-response) object
+ * or NodeJS [http.ServerResponse](https://nodejs.org/api/http.html#http_class_http_serverresponse) object
+ * @param {Object} [additionalFields] Object containing additional fields to be added.
+
+ * @return {Object}
+ * @ignore
+ */
+const createRequestDebugInfo = (request, response = {}, additionalFields = {}) => {
+    checkParamOrThrow(request, 'request', 'Object');
+    checkParamOrThrow(response, 'response', 'Object');
+    checkParamOrThrow(additionalFields, 'additionalFields', 'Object');
+
+    return Object.assign(
+        {
+            requestId: request.id,
+            url: request.url,
+            method: request.method,
+            retryCount: request.retryCount,
+            errorMessages: request.errorMessages,
+            // Puppeteer response has .status() funtion and NodeJS response ,statusCode property.
+            statusCode: _.isFunction(response.status) ? response.status() : response.statusCode,
+        },
+        additionalFields,
+    );
+};
 
 /**
  * A namespace that contains various utilities.
@@ -585,4 +629,5 @@ export const publicUtils = {
     htmlToText,
     URL_NO_COMMAS_REGEX,
     URL_WITH_COMMAS_REGEX,
+    createRequestDebugInfo,
 };